Normalize gamePath on retro_load_game() (#81)
gamePath, which is set in retro_load_game(), is later assumed to be
normalized even though it actually isn't. If a path like
"C:/program.exe" is passed to retro_load_game() on Windows when dir is
extracted it will equal "C:/program.exe" instead of "C:/". This is
because std::string::find_last_of is looking for the Windows \ path
separator and not the / one.

Normalizing gamePath makes this work correctly.